requirements.txt
numpy
opencv-python
Pillow
cvlib
tensorflow
easyocr
imutils
